粗糙度，相机，

+++ RayTracing/hitable_list.cpp
@@ -1,7 +1,7 @@
 #include "hitable_list.h"
 
 
-bool hitable_list::hit(const ray& r, float t_min, float t_max, hit_record& rec) const
+bool hitable_list::hit(const ray& r, double t_min, double t_max, hit_record& rec) const
 {
 	hit_record temp_rec;
 	bool hit_anything = false;
